requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pymongo>=4.7.1
pdfplumber>=0.11.0
python-dotenv>=1.0.1
pydantic>=2.7.0
